test: fixing build dependencies (#19783)
Part of #9953
