+++ source/extensions/filters/http/on_demand/BUILD
@@ -30,11 +30,6 @@ envoy_cc_extension(
     name = "config",
     srcs = ["config.cc"],
     hdrs = ["config.h"],
-    # TODO(#9953) classify and clean up.
-    extra_visibility = [
-        "//test/common/access_log:__subpackages__",
-        "//test/integration:__subpackages__",
-    ],
     deps = [
         "//envoy/registry",
         "//source/extensions/filters/http/common:factory_base_lib",
