+++ tools/spelling/spelling_dictionary.txt
@@ -1012,6 +1012,7 @@ ratelimited
 ratelimiter
 rawseti
 rc
+rdsone
 readded
 readonly
 readv
